Refactoring features q/a add links
